switch to a PROJ 6.1.0 build
PROJ has changed quite a bit in the past year, largely as a result of plans laid out in the GDAL barn raising (gdalbarn.com). For the build this means a dependency on SQLite, since the CSV EPSG file is no more, and the data now lives in a SQLite database named `proj.db`.

I started this build mainly since in GDAL 3.0 PROJ 6+ is needed, and I want a GDAL 3 build. I see from https://proj4.org/news.html that this also removes `projects.h` and deprecates `proj_api.h` (a new API is introduced).

This commit only updates the build. Rewrapping the APIs with Clang.jl can be done separately. Also `generate_projection_codes.jl` is not rerun, since it no longer works, since it should make use of `proj.db`. Updating that can also be done separately.

This also removes REQUIRE and adds Project.toml. I raised the minimum julia version to 1.0, since 0.7 is past end of life, or are there any objections to this?